refactor: Update action builder component structure.
